Feature - Add basic text field for entering word to learn
