+++ wordLearning/NewWordCard/NewWordCardViewModel.swift
@@ -34,7 +34,7 @@ class NewWordCardViewModel: ObservableObject {
         _state = .init(initialValue: .inactive)
     }
     
-    func dragGestureChanged(_ dragValue: DragGesture.Value, offset: Double) {
+    func dragGestureChanged(_ dragValue: DragGesture.Value, fullDragDistance: Double) {
         switch state {
         case .inactive:
             state = .dragToActive
@@ -47,9 +47,9 @@ class NewWordCardViewModel: ObservableObject {
         let dragGestureYOffset = dragValue.translation.height
         
         if dragGestureYOffset < 0 && (state == .dragToActive) {
-            dragToCenterProgress = min(1.0, Double(abs(dragGestureYOffset)) / offset)
+            dragToCenterProgress = min(1.0, Double(abs(dragGestureYOffset)) / fullDragDistance)
         } else if dragGestureYOffset > 0 && (state == .dragToInactive) {
-            dragToCenterProgress = max(0.0, 1.0 - Double(dragGestureYOffset) / offset)
+            dragToCenterProgress = max(0.0, 1.0 - Double(dragGestureYOffset) / fullDragDistance)
         }
         
         backgroundOpacity = dragToCenterProgress
@@ -69,54 +69,8 @@ class NewWordCardViewModel: ObservableObject {
         }
     }
     
-    func changeStep(previous: Bool = false) {
-        switch state {
-        case .inactive:
-            state = previous ? .inactive : .dragToActive
-        case .dragToActive:
-            state = previous ? .inactive : .activeFront
-        case .dragToInactive:
-            state = previous ? .activeFront : .inactive
-        case .activeFront:
-            state = previous ? .dragToInactive : .activeBack
-        case .activeBack:
-            state = previous ? .activeFront : .saving
-        case .saving:
-            state = .inactive
-        }
-    }
-    
-    
-    func updateFrontViewModel() {
+    private func updateFrontViewModel() {
         frontViewModel.hintOpacity = max(0.0, (1.0 - 4 * dragToCenterProgress))
         frontViewModel.topActionsBlurRadius = (20 - 4 * (dragToCenterProgress * 20.0))
-    }
-}
-
-class NewWordCardFrontViewModel: ObservableObject {
-    let colors: [Color] = [.white, .orange, .red, .gray, .green, .blue]
-    
-    @Published var hintOpacity: Double
-    @Published var topActionsBlurRadius: Double
-
-    @Published var currentColor: Color
-    @Published var nextColor: Color
-    
-    init() {
-        _hintOpacity = .init(initialValue: 1.0)
-        _topActionsBlurRadius = .init(initialValue: 20.0)
-        _currentColor = .init(initialValue: colors[0])
-        _nextColor = .init(initialValue: colors[1])
-    }
-    
-    func selectNextColor() {
-        currentColor = calculateNextColor()
-        nextColor = calculateNextColor()
-    }
-    
-    private func calculateNextColor() -> Color {
-        let currentColorIndex = colors.firstIndex(of: currentColor) ?? 0
-        let nextColorIndex = (currentColorIndex + 1) % colors.count
-        return colors[nextColorIndex]
     }
 }
